requirements.txt
Django==1.8.4
Fabric==1.10.2
Pillow==2.9.0

bpython==0.14.2
django-authtools==1.2.0
django-debug-toolbar==1.3.2
django-extensions==1.5.6
django-localflavor==1.1
django-redis==4.2.0
djangorestframework==2.4.4
gunicorn==19.3.0
ipython[notebook]==4.0.0
pdbpp==0.8.2
pytest==2.7.2
pytest-django==2.8.0
psycopg2==2.6.1
pytz==2015.4
raven==5.5.0
redis==2.10.3
